+++ tests/00-runtests.js
@@ -6,9 +6,9 @@ var runCount = 0,
 	testCount = 0;
 
 [
- "./02-feed.js",
- "./03-events.js",
- "./05-stream.js"
+ "./01-events.js",
+ "./02-stream.js",
+ "./03-feed.js"
 ]
 .map(require)
 .forEach(function (test){
