+++ tests/02-stream.js
